# apps/studio-web/openapi/studio-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StudioApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/occurrences": {
      "get": {
        "tags": [
          "OpenCampus"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OccurrenceListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "OpenCampus"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOccurrenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceDetail"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/occurrences/{id}": {
      "get": {
        "tags": [
          "OpenCampus"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "OpenCampus"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveOccurrenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceValidationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceConflictOfOccurrenceDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOccurrenceRequest": {
        "required": [
          "operationId",
          "draft"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "draft": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OccurrenceDraft"
              }
            ]
          }
        }
      },
...
      "DraftProblem": {
        "required": [
          "path",
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "EditorRef": {
        "required": [
          "id",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "OccurrenceDetail": {
        "required": [
          "id",
          "rowVersion",
          "updatedAt",
          "updatedBy",
          "revisionId",
          "publication",
          "draft",
          "references"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int64"
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EditorRef"
              }
            ]
          },
          "revisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationSummary"
          },
          "draft": {
            "$ref": "#/components/schemas/OccurrenceDraft"
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotReference"
            }
          }
        }
      },
      "OccurrenceDraft": {
        "required": [
          "name",
          "sourceNote",
          "days"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcDayDraft"
            }
          }
        }
      },
...
      "OccurrenceListItem": {
        "required": [
          "id",
          "name",
          "days",
          "relatedEventCount",
          "publication",
          "updatedAt",
          "updatedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcDayItem"
            }
          },
          "relatedEventCount": {
            "type": "integer",
            "format": "int32"
          },
          "publication": {
            "type": "string"
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EditorRef"
              }
            ]
          }
        }
      },
...
      "OcDayDraft": {
        "required": [
          "id",
          "date",
          "publicStart",
          "publicEnd",
          "status",
          "cancelNote"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "publicStart": {
            "type": [
              "null",
              "string"
            ]
          },
          "publicEnd": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "cancelNote": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "OcDayItem": {
        "required": [
          "id",
          "date",
          "publicStart",
          "publicEnd",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "publicStart": {
            "type": [
              "null",
              "string"
            ]
          },
          "publicEnd": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "PublicationSummary": {
        "required": [
          "state",
          "publishedRevisionId",
          "publishedAt"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "publishedRevisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ReferenceConflictOfOccurrenceDetail": {
        "required": [
          "code",
          "title",
          "latest"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "latest": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OccurrenceDetail"
              }
            ]
          }
        }
      },
...
      "ReferenceValidationProblem": {
        "required": [
          "code",
          "title",
          "problems"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "problems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DraftProblem"
            }
          }
        }
      },
...
      "SaveOccurrenceRequest": {
        "required": [
          "operationId",
          "rowVersion",
          "draft"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int64"
          },
          "draft": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OccurrenceDraft"
              }
            ]
          }
        }
      },
...
      "SlotReference": {
        "required": [
          "eventId",
          "eventTitle",
          "slotId",
          "ocDayId",
          "timeMode",
          "start",
          "end",
          "slotStatus"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "slotId": {
            "type": "string",
            "format": "uuid"
          },
          "ocDayId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "timeMode": {
            "type": [
              "null",
              "string"
            ]
          },
          "start": {
            "type": [
              "null",
              "string"
            ]
          },
          "end": {
            "type": [
              "null",
              "string"
            ]
          },
          "slotStatus": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}